requirements.txt
keyboard==0.13.5
opencv_python==4.5.5.64
Pillow==9.4.0
PyAutoGUI==0.9.53
pytesseract==0.3.9
pywin32==303